Project Plan
Here is a draft of a Project Plan for this week. I have found templates others had used for this class (such as the graph) and add some more material that is relevant to our project. Feel free to edit or make your own iteration. It's strange because by the way the professor explained it, it seems like we'll all be submitting the same "Project Plan" in our own assignment section. I guess when were done we can take the docx file and add our own panache if we like (haha).