Phase I
Here is a draft for Phase I of our Sortie Logger. When you submit it you will have to zip the "Sorties" package to abide by the directions on the third page. Feel free as always to make revisions as you see fit, or reach out for questions. I will post it to the public discussion tomorrow as well as the source code.